move responses closer to handler
- make response SEP8 compliant
